// frontend/src/api/generated/model/lLMConfigUpdateProvider.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ZeroAI
 * OpenAPI spec version: 0.1.0
 */

export type LLMConfigUpdateProvider = typeof LLMConfigUpdateProvider[keyof typeof LLMConfigUpdateProvider];


export const LLMConfigUpdateProvider = {
  ollama: 'ollama',
  openai: 'openai',
} as const;